Reorganize next tests to add benchmarks
