+++ tests/next_tests/artifacts/dummy_package/dummy_module.py
